service-workers/service-worker/clients-get-cross-origin.https.html

<!DOCTYPE html>
<title>Service Worker: Clients.get across origins</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
let host_info = get_host_info();

const SCOPE = 'resources/clients-get-frame.html';
const SCRIPT = 'resources/clients-get-worker.js';
const other_origin_iframe =
  host_info['HTTPS_REMOTE_ORIGIN'] +
  base_path() +
  'resources/clients-get-cross-origin-frame.html';
// This test asserts the behavior of the Client API in cases where the client
// belongs to a foreign origin. It does this by creating an iframe with a
// foreign origin which connects to a server worker in the current origin.

promise_test(async t => {
  t.add_cleanup(async () => {
    if (frame1) frame1.remove();
    if (frame2) frame2.remove();
    if (registration) await registration.unregister();
  });

  const registration = await service_worker_unregister_and_register(
    t,
    SCRIPT,
    SCOPE
  );
  await wait_for_state(t, registration.installing, 'activated');

  const frame1 = await with_iframe(SCOPE);
  let client_id = await new Promise(function(resolve, reject) {
    function get_client_id(e) {
      window.removeEventListener('message', get_client_id);
      resolve(e.data.clientId);
    }
    window.addEventListener('message', get_client_id, false);
  });

  const frame2 = await with_iframe(other_origin_iframe);
  frame2.contentWindow.postMessage(
    { clientId: client_id, type: 'getClientId' },
    host_info['HTTPS_REMOTE_ORIGIN']
  );
  client_id = await new Promise(function(resolve) {
    window.addEventListener('message', function(e) {
      if (e.data && e.data.type === 'clientId') {
        resolve(e.data.value);
      }
    });
  });
  assert_equals(client_id, undefined, 'iframe client ID');
}, 'Test Clients.get() cross origin');
</script>
